+++ projects/batfish-common-protocol/src/test/java/org/batfish/common/util/CollectionUtilTest.java
@@ -1,16 +1,21 @@
 package org.batfish.common.util;
 
 import static org.batfish.common.util.CollectionUtil.toImmutableMap;
+import static org.batfish.common.util.CollectionUtil.toOrderedHashCode;
+import static org.batfish.common.util.CollectionUtil.toUnorderedHashCode;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.Matchers.sameInstance;
 import static org.junit.Assert.assertThat;
 
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.ImmutableSortedMap;
 import com.google.common.collect.Maps;
 import java.util.Map;
+import java.util.stream.Stream;
 import org.junit.Test;
 
 /** Tests of {@link CollectionUtil}. */
@@ -33,5 +38,23 @@ public class CollectionUtilTest {
 
     assertThat(toImmutableMap(m), not(sameInstance(m)));
     assertThat(toImmutableMap(m), instanceOf(ImmutableMap.class));
+  }
+
+  @Test
+  public void testToOrderedHashCode() {
+    assertThat(Stream.of().collect(toOrderedHashCode()), equalTo(ImmutableList.of().hashCode()));
+    assertThat(Stream.of(1).collect(toOrderedHashCode()), equalTo(ImmutableList.of(1).hashCode()));
+    assertThat(
+        Stream.of(1, 2).collect(toOrderedHashCode()), equalTo(ImmutableList.of(1, 2).hashCode()));
+    assertThat(
+        Stream.of(2, 1).collect(toOrderedHashCode()), equalTo(ImmutableList.of(2, 1).hashCode()));
+  }
+
+  @Test
+  public void testToUnorderedHashCode() {
+    assertThat(Stream.of().collect(toUnorderedHashCode()), equalTo(ImmutableSet.of().hashCode()));
+    assertThat(Stream.of(1).collect(toUnorderedHashCode()), equalTo(ImmutableSet.of(1).hashCode()));
+    assertThat(
+        Stream.of(1, 2).collect(toUnorderedHashCode()), equalTo(ImmutableSet.of(2, 1).hashCode()));
   }
 }
